backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.5.8",
    "fastapi>=0.136.0",
    "joblib>=1.5.3",
    "openai>=2.32.0",
    "pandas>=3.0.2",
    "pydantic-settings>=2.14.0",
    "sentence-transformers>=5.4.1",
    "uvicorn[standard]>=0.44.0",
    "vadersentiment>=3.3.2",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.3",
]
